// src/app/Services/shared/enums/operation-type.ts
export enum OperationType {
  transAlExteriorConCambioD = 7,
  pagoHaberes = 8,
  pagoDeServicios = 9,
  pagoDeServiciosTelefonía = 10,
  consultarCuentas = 12,
  transferenciasCuentasPropias = 13,
  transferenciasCuentasTerceros = 14,
  nationalTransfers = 15,
  pagoProveedores = 16,
  formularioSolicitud = 17,
  boletaGarantia = 18,
  identificacionDeAbonos = 19,
  pagoMultiple = 20,
  pagoCreditoEmpresarial = 21,
  pagoFavorito = 23,
  pagoProveedoresAch = 24,
  pagoProveedoresEfe = 25,
  pagoProveedoresChequeGerencia = 26,
  pagoProveedoresOtrosBancosCheque = 27,
  pagoImpuestoGerencia = 28,
  ordenesDeDebitoAOtrosBancos = 29,
  formularioModificacion = 30,
  pagoServicioElfec = 32,
  pagoServicioPase = 33,
  pagoServicioEntel = 37,
  pagoServicioTigo = 38,
  pagoServicioNewPase = 40,
  pagoProveedoresEfectivoOnLine = 41,
  pagoQr = 44,
  pagoSoli = 45,
  pagoProveedoresChequeGerenciaLip = 46,
}